requirements.txt
# Core dependencies
geckoterminal-py>=0.2.5
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
pyyaml>=6.0
watchdog>=3.0.0

# Async support
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0
aiohttp>=3.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Optional database drivers
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.1.0  # MySQL

# QLib integration (optional)
# qlib>=0.9.0  # Uncomment if QLib integration is needed